Configure seed-job to poll hourly
To make it easier to make changes to the Jenkins jobs, this patch
configures the seed-job to poll the openzfs-ci repository hourly, such
that any changes to the repository will trigger the jobs to get rebuilt.
This mostly removes the need to manually run the seed-job via the
Jenkins UI whenever changes are made to the Jenkins jobs.