Use an empty string for PERL_PKGVERS
It turns out, the "SUNWcs" package fails to install into the new root
filesystem created when executing "onu" due to the following conflict:

    Creating Plan (Solver setup): -
    pkg install: No matching version of SUNWcs can be installed:
      Reject:  pkg://on-nightly/SUNWcs@0.5.11,5.11-0.151018:20161101T185932Z
      Reason:  A version for 'require' dependency on pkg:/runtime/perl-5161 cannot be found

When using the value of "-5161" for PERL_PKGVERS, the perl packages that
get created by the nigtly build system will have this value inserted
into the packages' names; e.g. "runtime/perl-5161/module/sun-solaris"
and "runtime/perl-5161". As a result, the "SUNWcs" package will have a
dependency on these perl packages with the "-5161" value in the package
name.

The problem occurs when these new perl packages are attempted to be
installed into the root filesystem created by "onu". The OmniOS system
that is used to build and perform the installation has the same perl
packages installed, but these package names do not contain the "-5161"
portion. Thus, when we run "onu", the perl package already installed in
the system will get pulled into the new root image. Then when the new
packages created by the nightly build are attempted to be installed,
they will conflict with the perl packages already installed, causing the
installation to fail. Additionally, since the newly built "SUNWcs"
package depends on the perl package that failed to install, this package
will also fail to install; this will result in the "SUNWcs" package from
the omnios publisher being installed in the new root filesystem, rather
than the "SUNWcs" package that was built via the nightly build.

To fix this issue, we're using an empty string as the value of
PERL_PKGVERS such that the nightly build will create perl packages that
match the names of what the OmniOS build system is already using. Thus,
these newly built packages will replace the older, already installed,
packages instead of conflicting with them.