Don't overwrite existing test results
This patch modifies the URL and directory structure used when storing
test results. Previously if a pull request or commit was ever re-tested,
any previous results would be overwritten by the results of the most
recent test iteration. This change modifies this behavior such that we
preserve the results of all test iterations performed for any given pull
request or commit.

Closes #4
Closes #5