Automate the sending of mail to illumos developers
This change implements a new Jenkins job whose purpose is to automate
the process of sending mail to the illumos developer mailing list. As
per the illumos integration requirements, when a pull request is opened
isn't specific to OpenZFS, we need to send a mail to the illumos
developer mailing list in order for that change to be accepted by the
illumos advocates (remember, all changes that integrate into the OpenZFS
repository, are also integrated into the illumos repository).

Without this change, the burden of creating and sending this mail is
placed on the author of the pull request, who may or may not be familiar
with this process. Additionally, it's not clearly stated in the pull
request's history when the mail was sent; or even _if_ the mail was
sent.

With this change, a label can be added to a given pull request to
indicate that a mail needs to be sent to the illumos developer list.
The automated job will eventually detect this label, create the mail
required, and send it to the appropriate list. Additionally, the owner
of the pull request, along with the authors and committers of any
commits associated with the pull request will be CC'd on the email.

After the mail is sent, a label will be automatically added to the pull
request to indicate that the mail has been sent, and the original label
that indicated a mail needed to be sent, will be removed.

Closes #3
Closes #7