Automate illumos sync pull requests
This change implements a new Jenkins job whose purpose is to automate
the process of merging the illumos codebase with the OpenZFS codebase,
and opening a new pull request with the merged code against the OpenZFS
project.

If this job runs and finds a pull request already open for this
"illumos sync" purpose, a new one will not be opened; instead, the
automation will wait for the currently open pull request to be closed,
before opening a new one.

Closes #10