Don't automatically restart docker containers
When running the ansible role, if the docker image for the containers
was modified, the container would be restarted so that it would running
using the most up-to-date version of the docker image. If there happened
to be any regressions tests in-progress when the container restarted,
these jobs would fail.

To prevent these failures, the "ignore_image" option is being used to
prevent these automatic container restarts. Now, to have the container
start using the new docker image, the container will have to be manually
updated.

Also, since the "ignore_image" option is only available when using
Ansible v2.2, the script that pulls ansible into the virtual environment
was updated to require version 2.2.