Use the "user" ansible module
If the "jenkins" user did not previously exist, the "jenkins-slave"
module would fail rather than creating the user. The assumption was for
this user to be provided by LDAP, so we didn't need to worry about
creating the user, since it should always exist.

To support new systems that don't have LDAP enabled, we need to support
the case where the "jenkins" user doesn't already exist, and instead of
failing, the "jenkins" user should be added to the system's local users.