Stop using Jenkins Swarm plugin
This commits marks the end of using the Jenkins Swarm plugin for
connecting dynamic slaves to the Jenkins master, and opts for using the
more integrated way for connecting slaves described here:

    https://wiki.jenkins-ci.org/display/JENKINS/Distributed+builds

The motivating factor for this change was the fact that "connection
reset" errors were seen when using the Jenkins Swarm plugin on
OpenIndiana based slaves, but these errors were not seen when using the
"native" JNLP connection (which this change implements). Thus, with this
in place, it should be easier to start using OpenIndiana based slaves.

In addition to the above, the mechanism for executing the client's slave
process was migrated from an SMF service, to using supervisord. When
using SMF with the new JNLP based slave, the slave's connection was
continually being terminated, but when using supervisord, these issue
didn't reproduce. So while the root cause for the terminated connections
isn't clear, it appears to have been "fixed" by the migration to using
supervisord. The added benefit of this, is supervisord should be
portable to other platform and operating systems (e.g. linux), whereas
the SMF service would have had to be re-implemented to use the new
platform's init service.

And finally, lots of the Ansible role's variables were renamed and
refactored, hopefully to be more clear and representative of what a
consumer of the role might actually want to define and/or modify. For
example, the "jenkins_master" was renamed to "jenkins_master_url" to be
more descriptive, and it's not possible to easily specific which java
binary to use via the new "java_bin_path" variable.