Refactor build script and Ansible role
This change refactors the "openzfs.build-slave" Ansible role to use the
metapackage available in OmniOS to pull down all of the necessary tools
needed to build the OpenZFS and illumos repositories. This simplifies
the role, and is less prone to errors as that metapackage is more likely
to be correct than the list of packages previously used.

Also, the role was updated to accomodate OpenIndiana, such that if the
role is applied to an OpenIndiana based system, the correct packages
will be installed for building OpenZFS and illumos on OpenIndiana.

Finally, the "nightly-build.sh" script was also simplified and updated
to accomodate building on OpenIndiana. Specifically, the "illumos.sh"
environment file provided by OmniOS is now leaned on, instead of
manually specifying the nightly build options necessary for building on
OmniOS; again, this is more likely to be correct than our previous
method of configuring the environment file. When not building on OmniOS,
the "illumos.sh" contained in the repository is used, but only minimal
changes are made (since the default values work well for building on
OpenIndiana).