Automation for generating custom OI install media
This patch adds a new Jenkins job and accompanying scripts for
generating OpenIndiana install media using a custom build of illumos or
OpenZFS.

The automation takes the OpenZFS (or illumos) repository and branch as a
parameter, and will then proceed to perform a full nightly build of that
specific branch, and will then build the install media using that custom
OpenZFS/illumos build.

The contents of the "/rpool/dc/media" directory will then be uploaded to
Joyent's Manta, allowing relatively easy access to the ".iso" and ".usb"
artifacts of "distro_const" using simple tools such as "wget" and "tar".

For example, the following commands can be used to download and extrace
the install media generated by this automation for the OpenZFS
repository's "master" branch:

    $ wget http://us-east.manta.joyent.com/zettabot/public/install-media/openzfs/openzfs/master/install-media.tar.xz
    $ tar -xf install-media.tar.xz
    $ ls -lh install-media.tar.xz OpenIndiana_Text_X86.*
    -rw-r--r-- 1 root root 1.3G Feb 16 14:22 install-media.tar.xz
    -rw-r--r-- 1 root root 649M Feb 16 14:07 OpenIndiana_Text_X86.iso
    -r--r--r-- 1 root root 783M Feb 16 14:07 OpenIndiana_Text_X86.usb

The format for the URL is

    http://us-east.manta.joyent.com/zettabot/public/install-media/$USER/$REPO/$BRANCH/install-media.tar.xz

where:

  - `$USER` corresponds to the GitHub user or organization for the given
    openzfs/illumos repository that will be built and used in the
    installation media. E.g. "openzfs", "illumos", etc.

  - `$REPO` corresponds to the name of the GitHub repository that
    contains the openzfs/illumos code that will be built. E.g.
    "openzfs", "illumos-gate", etc.

  - `$BRANCH` corresponds to the git branch of the openzfs/illumos
    repository that will be used when performing the "nightly" build.