Fix result tarballs for zfstest and zloop
As of the following commit:

    commit bd6e151d99528fcaaea38ca493cf62c41fb466b3
    Author: Prakash Surya <prakash.surya@delphix.com>
    Date:   Wed Feb 8 20:18:01 2017 -0800

        Refactor build script and Ansible role

the tarballs that are generated from zfstest and zloop runs have been
broken due to the OmniOS build system not having the "gtar" utility
installed. This patch fixes the issue by adding the necessary package to
the list of packages installed when configuring the build system.