Migrate CI server to Joyent's public cloud
This change migrates all the CI server configuration to work on the new
VM deployed on Joyent's public cloud. In addition to migrating to
Joyent's cloud, the various Jenkins job DSL files are now being stored
directly in the OpenZFS repository, so the "seed-job" now uses that
repository to generate the various jobs instead of this repository.